wip: use best_route from within windows
