+++ rust/socket-factory/Cargo.toml
@@ -13,12 +13,3 @@ tracing = "0.1"
 
 [features]
 hickory = ["dep:hickory-proto", "dep:async-trait"]
-
-[target.'cfg(target_os = "windows")'.dependencies.windows]
-version = "0.57.0"
-features = [
-  "Win32_Foundation",
-  "Win32_NetworkManagement_IpHelper",
-  "Win32_NetworkManagement_Ndis",
-  "Win32_Networking_WinSock",
-]
