+++ rust/socket-factory/src/lib.rs
@@ -1,9 +1,9 @@
-use std::net::{IpAddr, SocketAddr};
+use std::collections::HashMap;
 use std::{
     borrow::Cow,
     collections::VecDeque,
     io::{self, IoSliceMut},
-    net::SocketAddr,
+    net::{IpAddr, SocketAddr},
     slice,
     task::{ready, Context, Poll},
 };
@@ -72,6 +72,8 @@ impl std::os::fd::AsFd for TcpSocket {
 pub struct UdpSocket {
     inner: tokio::net::UdpSocket,
     state: quinn_udp::UdpSocketState,
+    source_ip_resolver: Box<dyn Fn(IpAddr) -> Option<IpAddr> + Send + Sync + 'static>,
+    routes: HashMap<IpAddr, IpAddr>,
 
     port: u16,
 
@@ -86,8 +88,17 @@ impl UdpSocket {
             state: quinn_udp::UdpSocketState::new(quinn_udp::UdpSockRef::from(&inner))?,
             port,
             inner,
+            source_ip_resolver: Box::new(|_| None),
             buffered_datagrams: VecDeque::new(),
+            routes: Default::default(),
         })
+    }
+
+    pub fn set_source_ip_resolver(
+        &mut self,
+        resolver: Box<dyn Fn(IpAddr) -> Option<IpAddr> + 'static + Send + Sync>,
+    ) {
+        self.source_ip_resolver = resolver;
     }
 }
 
@@ -225,13 +236,25 @@ impl UdpSocket {
         }
     }
 
-    pub fn try_send(&self, transmit: &DatagramOut) -> io::Result<()> {
+    fn get_source(&mut self, dst: IpAddr) -> Option<IpAddr> {
+        match self.routes.entry(dst) {
+            std::collections::hash_map::Entry::Occupied(entry) => Some(*entry.get()),
+            std::collections::hash_map::Entry::Vacant(v) => {
+                let src = (self.source_ip_resolver)(dst)?;
+                Some(*v.insert(src))
+            }
+        }
+    }
+    pub fn try_send(&mut self, transmit: &DatagramOut) -> io::Result<()> {
         let transmit = quinn_udp::Transmit {
             destination: transmit.dst,
             ecn: None,
             contents: &transmit.packet,
             segment_size: None,
-            src_ip: transmit.src.map(|s| s.ip()),
+            src_ip: transmit
+                .src
+                .map(|s| s.ip())
+                .or_else(|| self.get_source(transmit.dst.ip())),
         };
 
         self.inner.try_io(Interest::WRITABLE, || {
@@ -296,86 +319,5 @@ mod hickory {
         ) -> Poll<io::Result<usize>> {
             <TokioUdpSocket as DnsUdpSocketTrait>::poll_send_to(&self.inner, cx, buf, target)
         }
-    }
-}
-
-#[cfg(target_os = "windows")]
-fn get_best_route(dst: IpAddr) -> IpAddr {
-    use std::mem::{size_of, size_of_val, MaybeUninit};
-    use std::net::{Ipv4Addr, Ipv6Addr};
-    use std::ptr::null;
-
-    use windows::Win32::NetworkManagement::IpHelper::GetAdaptersAddresses;
-    use windows::Win32::NetworkManagement::IpHelper::GetBestRoute2;
-    use windows::Win32::NetworkManagement::IpHelper::GET_ADAPTERS_ADDRESSES_FLAGS;
-    use windows::Win32::NetworkManagement::IpHelper::IP_ADAPTER_ADDRESSES_LH;
-    use windows::Win32::NetworkManagement::IpHelper::MIB_IPFORWARD_ROW2;
-    use windows::Win32::Networking::WinSock::ADDRESS_FAMILY;
-    use windows::Win32::Networking::WinSock::AF_UNSPEC;
-    use windows::Win32::Networking::WinSock::SOCKADDR_INET;
-    // SAFETY: lol
-    unsafe {
-        // TODO: iterate until it doesn't overflow
-        let mut addresses: Vec<u8> = vec![0u8; 15000];
-        let mut addresses_len = size_of_val(&addresses) as u32;
-        GetAdaptersAddresses(
-            AF_UNSPEC.0 as u32,
-            GET_ADAPTERS_ADDRESSES_FLAGS(0),
-            Some(null()),
-            Some(addresses.as_mut_ptr() as *mut _),
-            &mut addresses_len as *mut _,
-        );
-
-        let mut next_address = addresses.as_ptr() as *const _;
-        let mut luids = Vec::new();
-        loop {
-            let address: &IP_ADAPTER_ADDRESSES_LH = std::mem::transmute(next_address);
-            luids.push(address.Luid);
-            if address.Next.is_null() {
-                break;
-            }
-            next_address = address.Next;
-        }
-
-        let mut routes: Vec<(MIB_IPFORWARD_ROW2, SOCKADDR_INET)> = Vec::new();
-        for luid in &luids {
-            let addr: SOCKADDR_INET = SocketAddr::from((dst, 0)).into();
-            let mut best_route: MaybeUninit<MIB_IPFORWARD_ROW2> = MaybeUninit::zeroed();
-            let mut best_src: MaybeUninit<SOCKADDR_INET> = MaybeUninit::zeroed();
-
-            GetBestRoute2(
-                Some(luid as *const _),
-                0,
-                None,
-                &addr as *const _,
-                0,
-                best_route.as_mut_ptr(),
-                best_src.as_mut_ptr(),
-            );
-
-            let best_route = best_route.assume_init();
-            let best_src = best_src.assume_init();
-            routes.push((best_route, best_src));
-        }
-
-        routes.sort_by(|(a, _), (b, _)| a.Metric.cmp(&b.Metric));
-
-        let addr = routes.first().unwrap().1;
-        match addr.si_family {
-            ADDRESS_FAMILY(0) => todo!(),
-            ADDRESS_FAMILY(2) => Ipv4Addr::from(addr.Ipv4.sin_addr).into(),
-            ADDRESS_FAMILY(23) => Ipv6Addr::from(addr.Ipv6.sin6_addr).into(),
-            _ => panic!("Invalid address"),
-        }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-    #[cfg(target_os = "windows")]
-    #[test]
-    fn best_route_works() {
-        dbg!(get_best_route("8.8.8.8".parse().unwrap()));
     }
 }
